arm: dts: imx8mp-data-modul-edm-sbc: Remove deprecated sd-vsel-gpios
The sd-vsel-gpios property in the root of the PMIC node is deprecated
and therefore not parsed by the driver anymore. We can safely remove
this as it wasn't used anyway due to the pad not having the correct
pinmux settings.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>
Reviewed-by: Marek Vasut <marek.vasut@mailbox.org>
Signed-off-by: Peng Fan <peng.fan@nxp.com>
(cherry picked from commit 925f63b0209483cd04b014cc8b5deab1f20267b7)